Don't assume that Host is set in ssh/config
SSH does not require "Hostname" to be set, in which case the "Host" line
denotes the Hostname, as in:

Host github.com
  Compression yes
  User git
  IdentityFile %d/.ssh/github

magit-gh-pulls-get-ssh-config-hosts does not set the cadr of the parsed
host name cell in this case, and consequently magit-gh-pulls-parse-url
failed to get the correct hostname in this case.

Now we check whether the host name is actually configured, and only
update the host name if it's there.