Replace deprecated magit-section-value
Hi there! :wave: Big fan of the project!

Problem
=======

As of Magit 2.12.0, the function magit-section-value became obsolete. As of recently, the function no longer exists in Magit. Consequently, the functions `magit-gh-section-req-data` and `magit-gh-pulls-open-in-browser` are throwing an error for me.

Solution
========

As recommended by Magit, I have replaced the `(magit-section-value ...)` calls with `(oref ... value)`. Works on my machine!